requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
tensorflow-cpu==2.11.0
numpy
scipy
librosa
soundfile
